bot/requirements.txt
discord.py>=2.3.0
asyncpg>=0.29.0
python-socketio[asyncio_client]==5.11.1
python-engineio==4.9.0
python-dotenv>=1.0.1
aiohttp==3.9.3
